delete data - OK statisticsServiceImpl - TODO

+++ src/main/java/searchengine/dto/statistics/TotalStatistics.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class TotalStatistics {
     private int sites;
-    private int pages;
-    private int lemmas;
+    private long pages;
+    private long lemmas;
     private boolean indexing;
 }
